# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCredentials {
  username: string;
  password: string;
}