# lib/api-zod/src/generated/types/pageContent.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface PageContent {
  id: number;
  page: string;
  sectionKey: string;
  valueEn: string;
  valuePl: string;
  valueRu: string;
  updatedAt?: string;
}