# lib/api-zod/src/generated/types/popularVehicleInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface PopularVehicleInput {
  make: string;
  model: string;
  imageUrl?: string;
  priceRange: string;
  estimatedDelivery: string;
  description: string;
  sortOrder?: number;
}