# lib/api-zod/src/generated/types/pricingPackageInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface PricingPackageInput {
  nameEn: string;
  namePl?: string;
  nameRu?: string;
  price: number;
  currency?: string;
  features?: string[];
  isPopular?: boolean;
  sortOrder?: number;
}