# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  iconName: string;
  titleEn: string;
  titlePl?: string;
  titleRu?: string;
  descriptionEn: string;
  descriptionPl?: string;
  descriptionRu?: string;
  sortOrder: number;
  createdAt?: string;
}