# lib/api-zod/src/generated/types/siteSetting.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSetting {
  id: number;
  key: string;
  value: string;
  label?: string;
  group?: string;
  updatedAt?: string;
}