# lib/api-zod/src/generated/types/soldVehicleInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface SoldVehicleInput {
  make: string;
  model: string;
  year: number;
  /** @nullable */
  finalPrice?: number | null;
  purchaseCountry: string;
  deliveryStatus?: string;
  /** @nullable */
  deliveryDate?: string | null;
  imageUrl?: string;
}